Renamed target and files from QF-QM to QPC-QM
