+++ comodo2/src/comodo2/engine/Config.java
@@ -27,7 +27,7 @@ public class Config {
 	public final static  String TARGET_PLATFORM_SCXML   = "SCXML";      // SCXML only
 	public final static  String TARGET_PLATFORM_ELT_RAD = "ELT-RAD";    // RAD application (SCXML, MAL, rad.ev)
 	public final static  String TARGET_PLATFORM_ELT_MAL = "ELT-MAL";    // MAL ICD only
-	public final static  String TARGET_PLATFORM_QF_QM = "QF-QM";    	// Quantum Framework: Quantum Modeler file format
+	public final static  String TARGET_PLATFORM_QPC_QM = "QPC-QM";    	// Quantum Framework: Quantum Modeler file format
 
 	/**
 	 * Platforms specific configuration options
