+++ comodo2/src/comodo2/templates/qpc/qm/QmTree.java
@@ -1,4 +1,4 @@
-package comodo2.templates.qf.qm;
+package comodo2.templates.qpc.qm;
 
 import java.util.ArrayList;
 import java.util.Objects;
